requirements.txt
anthropic>=0.25.0
openai>=1.0.0
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pypdf>=4.0.0
tiktoken>=0.6.0
pandas>=2.0.0
numpy>=1.24.0
